+++ datajoint/autopopulate.py
@@ -27,12 +27,12 @@ class AutoPopulate:
     @property
     def key_source(self):
         """
-        :return: the relation whose primary key values are passed, sequentially, to the
-                ``make`` method when populate() is called.
+        :return: the query whose primary key values are passed, sequentially, to the
+                `make` method when populate() is called.
                 The default value is the join of the parent relations.
                 Users may override to change the granularity or the scope of populate() calls.
         """
-        def parent_gen(self):
+        def parent_gen():
             if self.target.full_table_name not in self.connection.dependencies:
                 self.connection.dependencies.load()
             for parent_name, fk_props in self.target.parents(primary=True).items():
@@ -44,7 +44,7 @@ class AutoPopulate:
                         attr: ref for attr, ref in fk_props['attr_map'].items() if ref != attr})
 
         if self._key_source is None:
-            parents = parent_gen(self)
+            parents = parent_gen()
             try:
                 self._key_source = next(parents)
             except StopIteration:
@@ -64,8 +64,8 @@ class AutoPopulate:
     @property
     def target(self):
         """
-        relation to be populated.
-        Typically, AutoPopulate are mixed into a Relation object and the target is self.
+        :return: table to be populated.
+        In the typical case, dj.AutoPopulate is mixed into a dj.Table class by inheritance and the target is self.
         """
         return self
 
@@ -73,6 +73,7 @@ class AutoPopulate:
         """
         :param key:  they key returned for the job from the key source
         :return: the dict to use to generate the job reservation hash
+        This method allows subclasses to control the job reservation granularity.
         """
         return key
 
@@ -142,7 +143,7 @@ class AutoPopulate:
 
         make = self._make_tuples if hasattr(self, '_make_tuples') else self.make
 
-        for key in (tqdm(keys) if display_progress else keys):
+        for key in (tqdm(keys, desc=self.__class__.__name__) if display_progress else keys):
             if max_calls is not None and call_count >= max_calls:
                 break
             if not reserve_jobs or jobs.reserve(self.target.table_name, self._job_key(key)):